+++ Wire-iOS/Sources/UserInterface/ConversationList/ConversationListViewController+StartUI.swift
@@ -25,39 +25,3 @@ extension ConversationListViewController {
         return startUIViewController
     }
 }
-
-///TODO: new file
-extension ConversationListViewController {
-    @objc
-    func setState(_ state: ConversationListState, animated: Bool, completion: (() -> ())?) {
-        if self.state == state {
-            completion?()
-            return
-        }
-        self.state = state
-
-        switch state {
-        case .conversationList:
-            view.alpha = 1
-
-            if let presentedViewController = presentedViewController {
-                presentedViewController.dismiss(animated: true, completion: completion)
-            } else {
-                completion?()
-            }
-        case .peoplePicker:
-            let startUIViewController = createPeoplePickerController()
-            let navigationWrapper = startUIViewController.wrapInNavigationController(ClearBackgroundNavigationController.self)
-
-            show(navigationWrapper, animated: true) {
-                startUIViewController.showKeyboardIfNeeded()
-                completion?()
-            }
-        case .archived:
-            show(createArchivedListViewController(), animated: animated, completion: completion)
-        @unknown default:
-            break
-        }
-    }
-
-}
