+++ Wire-iOS/Sources/UserInterface/ConversationList/UIViewControllerTransitioningDelegate+Present.swift
